back/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
mysqlclient==2.0.3
Pillow==8.3.1
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
